Move relayer to subfolder along with docs update
