+++ tests/relay_message_api.go
@@ -14,7 +14,7 @@ import (
 
 	"github.com/ava-labs/avalanchego/ids"
 	avalancheWarp "github.com/ava-labs/avalanchego/vms/platformvm/warp"
-	"github.com/ava-labs/awm-relayer/api"
+	"github.com/ava-labs/awm-relayer/relayer/api"
 	testUtils "github.com/ava-labs/awm-relayer/tests/utils"
 	"github.com/ava-labs/subnet-evm/core/types"
 	subnetEvmInterfaces "github.com/ava-labs/subnet-evm/interfaces"
